requirements.txt
boto==2.43.0
paho-mqtt==1.2
